Refactor label generation functions into utils.plotting submodule
Refactor label generation functions into simpler structure
